Role and RoleBinding objects are namespaced.
It was recently discovered that some of our deployments were
broken due to Role and RoleBinding objects not being
resolvable in clouddriver. This appears to be due to a fix in
clouddriver outlined in #2984. The fix will prevent caching
of cluster-scoped objects in a namespace.

Since we have marked Role and RoleBinding as not namespaced,
they are being ignored by the caching agents, as outlined in
the pull request above.

This PR marks the resources as namespaced, as it is outlined
in the k8s API documentation that they can be. This should
mean the resources will pass the check in #2984,  allowing
them to be cached going forward.