Statefulset Partition roll out support
This PR adds support for statefulset partition roll outs. In
order to do this with some confidence, we've added specs
around the `KubernetesStatefulSetHandler` as well.

During this process a small bug was uncovered with
`KubernetesManifest.isNewerThanObservedGeneration()`.

The `generation` and `observedGeneration` properties of
`KubernetesManifest` are hydrated from YAML as integers at
some point, so you end up with a `ClassCastException`, and
that was being ignored, so the method ultimately returned
true.

Note that these properties are integers according to the
kubernetes API docs.

https://kubernetes.io/docs/reference/generated/kubernetes-api/v1.10/#objectmeta-v1-meta

https://kubernetes.io/docs/reference/generated/kubernetes-api/v1.10/#rollingupdatestatefulsetstrategy-v1-apps

The logic was taken directly from the kubectl code here
https://github.com/kubernetes/kubernetes/blob/master/pkg/kubectl/rollout_status.go#L137-L146
c922c2f

Looking for input on making sure I've covered everything.